commit to main

+++ base/views.py
@@ -1,5 +1,10 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
 from .models import Room, Topic
 from .forms import RoomForm
 #from django.http import HttpResponse
@@ -12,6 +17,36 @@ from .forms import RoomForm
 #     {'id':3 , 'name':'frontend developer'},
 # 
 # ]
+
+def loginPage(request):
+
+    if request.method =='POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+
+        try: 
+            user = User.objects.get(username=username)
+        except: 
+            messages.error(request, 'User does not exist')
+             
+        user =authenticate(request,username = username, password= password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Username or Password does not exist')
+
+    context= {}
+    return render(request, 'base/login_register.html', context)
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+    
+    
+
 def home(request):
     q = request.GET.get('q') if request.GET.get('q')!= None else ''
 
@@ -37,6 +72,7 @@ def room(request,pk):
     context = {'room':room}
     return render(request,'base/room.html',context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form =  RoomForm()
     if request.method == 'POST':
@@ -48,9 +84,13 @@ def createRoom(request):
     return render(request, 'base/room_form.html',context)
 
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance = room)
+
+    if request.user != room.host: 
+        return HttpResponse('You are not aloowed here')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -59,9 +99,15 @@ def updateRoom(request, pk):
             return redirect('home')
     context= {'form':form}
     return render(request, 'base/room_form.html',context)
-    
+
+
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host: 
+        return HttpResponse('You are not aloowed here')
+        
     if request.method == 'POST':
         room.delete()
         return redirect('home')
